Made some relatively substantial changes
1. Added new multiplication based hashing algorithm and modified
the hash class to support it. I added:

	a private method to calculate appropriate multiplication parameters
	a vector to store these parameters

This relies on using rand() though so I don't know if it produces intelligent enough
choices or not.

2. Changed the driver slightly. It now takes 3 command line arguments:

./cuckoo <input> <output> {M|D}

M signals to use the multiplication method and D signals to use the division method.

----

If I feel up to it I'll also add in the prime number system I brought up briefly in class.